Refactor dragon engine
