+++ core/modes/modes_dragon.talon
@@ -23,4 +23,4 @@ speech.engine: dragon
     user.help_hide()
     user.mouse_sleep()
     speech.disable()
-    user.engine_sleep()
+    user.dragon_engine_sleep()
